smi: Update mainboard_smi_gpi() to have 32bit argument
With the LynxPoint chipset there are more than 16
possible GPIOs that can trigger an SMI so we need
a mainboard handler that can support this.

There are only a handful of users of this function
so just change them all to use the new prototype.

Reviewed-on: https://gerrit.chromium.org/gerrit/49530
Reviewed-on: http://review.coreboot.org/4145
Tested-by: build bot (Jenkins)

+++ src/mainboard/getac/p470/smihandler.c
@@ -180,7 +180,7 @@ static void mainboard_smi_hotkey(u8 hotkey)
 	printk(BIOS_DEBUG, "EC hotkey: %02x\n", hotkey);
 }
 
-void mainboard_smi_gpi(u16 gpi_sts)
+void mainboard_smi_gpi(u32 gpi_sts)
 {
 	u8 source, hotkey;
 	send_ec_oem_command(0x5c);
